Move PriceTagStatus to proof constants
