+++ open_prices/proofs/constants.py
@@ -1,3 +1,5 @@
+import enum
+
 TYPE_PRICE_TAG = "PRICE_TAG"
 TYPE_RECEIPT = "RECEIPT"
 TYPE_GDPR_REQUEST = "GDPR_REQUEST"
@@ -31,3 +33,13 @@ PRICE_TAG_EXTRACTION_TYPE = "PRICE_TAG_EXTRACTION"
 PRICE_TAG_PREDICTION_TYPE_CHOICES = [
     (PRICE_TAG_EXTRACTION_TYPE, PRICE_TAG_EXTRACTION_TYPE)
 ]
+
+
+class PriceTagStatus(enum.IntEnum):
+    deleted = 0
+    linked_to_price = 1
+    not_readable = 2
+    not_price_tag = 3
+
+
+PRICE_TAG_STATUS_CHOICES = [(item.value, item.name) for item in PriceTagStatus]
