+++ open_prices/stats/tests.py
@@ -1,7 +1,6 @@
 from django.db import IntegrityError
 from django.test import TestCase
 
-from open_prices.common.constants import PriceTagStatus
 from open_prices.locations.factories import LocationFactory
 from open_prices.prices import constants as price_constants
 from open_prices.prices.factories import PriceFactory
@@ -77,7 +76,7 @@ class TotalStatsTest(TestCase):
         PriceTagFactory(
             proof=cls.proof_price_tag,
             price=cls.price,
-            status=PriceTagStatus.linked_to_price.value,
+            status=proof_constants.PriceTagStatus.linked_to_price.value,
         )
 
     def test_update_price_stats(self):
